Moving the test file to it's correct place.
- tests moving around
- Changelog
